OMAPDSS: HDMI: Add function to get the HDMI mode
Add function to obtain the DVI/HDMI mode. This is needed by
the ASoC HDMI audio codec. HDMI audio can only be transmitted
when the IP is configured as HDMI. As the DVI/HDMI mode is a private
parameter of the HDMI driver, this functions allows to expose
such mode to the ASoC driver. Also, the function is exported for
dynamic linking as the ASoC codec and DSS may be built as separate
modules.

Signed-off-by: Ricardo Neri <ricardo.neri@ti.com>