OMAPDSS: Provide interface for audio support in DSS
There exist several display technologies and standards that support audio as
well. Hence, it is relevant to update the DSS device driver to provide an audio
interface that may be used by an audio or any other driver interested in the
functionality.

The audio_enable function is intended to prepare the relevant IP for playback
(e.g., enabling an audio FIFO, taking out of reset some IP, etc).

While the display may support audio, it is possible that for certain
configurations audio is not supported (e.g., an HDMI display using a VESA video
timing). The audio_detect function is intended to query whether the current
configuration of the display supports audio.

The audio_config function is intended to configure all the relevant audio
parameters of the display. In order to make the function independent of DSS
any device driver, it takes a IEC-60958 channel status word as well
as a CEA-861 audio infoframe. At the moment, this is enough to support HDMI
and DisplayPort, as both are based on CEA-861 and IEC-90658. Please note that
while asound.h is #included, the DSS device driver should not link to any
ALSA function as DSS and ALSA are built in separate modules.

The audio_start function is intended to effectively start/stop audio playback
after the configuration has taken place.

Signed-off-by: Ricardo Neri <ricardo.neri@ti.com>