OMAPDSS: Provide interface for audio support in DSS
There exist several display technologies and standards that support audio as
well. Hence, it is relevant to update DSS to provide an audio
interface to be used by the audio drivers.

At the moment, this interface is enough to support HDMI and, in the future,
DisplayPort, as bot are based on CEA-861 and IEC90658.

Signed-off-by: Ricardo Neri <ricardo.neri@ti.com>