OMAPDSS: HDMI: HACK: Set idle mode for HDMI audio
When playing audio, the idle mode of the HDMI module should be in
no-idle. This is because the HDMI audio wrapper uses the interface
clock to work, but it might be gated if not used by the HDMI IP core.

While there, also change the name of the argument to make it more
descriptive.

Signed-off-by: Ricardo Neri <ricardo.neri@ti.com>