OMAPDSS: HDMI: Remove ASoC codec
Remove the the ASoC OMAP HDMI audio codec. The operations performed by this
codec will be relocated to the ASoC HDMI CPU DAI driver under /sound/soc.
As such driver is in the audio domain, it will be submitted in a separate
patch.

This patch removes the current HDMI audio support. It will be restored in
subsequent patches using the DSS device driver audio interface. This will
make possible that any potential user, not only ASoC, may use the HDMI
audio functionality provided by DSS. Also, it will prevent dependency
issues if, for instance ASoC is compiled as a module and DSS is compiled
as built-in.

Signed-off-by: Ricardo Neri <ricardo.neri@ti.com>