OMAPDSS: HDMI: Implement DSS driver interface for audio
Implement the DSS device driver audio support interface in the HDMI
panel driver and generic driver. The implementation relies on the
IP-specific functions that are defined at DSS probe time.

A hardirq-safe spinlock is used to protect the audio_start/stop
functions.. This is because such function may be called while
holding a lock. The operation should not take too long to complete
as at this point all the required resources (DMA, audio FIFO,
regulators, companion chips, etc.) should be prepared to start
audio data transfers.

At the moment, the rest of the audio functions also are protected
by a hardirq-safe spinlock. In the future, a mutex may have to be
used, as the enablement of resources may take too much time.

Signed-off-by: Ricardo Neri <ricardo.neri@ti.com>