OMAPDSS: HDMI: Implement DSS driver interface for audio
Add support for audio in the HDMI panel driver using the DSS driver
audio interface.

A spinlock is used for the audio functions. This is because
the audio functions may be called while holding a lock; in such case,
the panel's driver is not suitable. Also, local interrupts are disabled
for increased security; functions should not take very long to return as
they mainly set registers.

Signed-off-by: Ricardo Neri <ricardo.neri@ti.com>