OMAPDSS: HDMI: Implement DSS driver interface for audio
Implement the DSS device driver audio interface in the HDMI panel and
driver.

A HW-safe spinlock is used to protect the audio functions. This is because
the audio functions may be called while holding a lock; in such case,
the panel's driver mutex is not suitable. Functions should be used
to set registers and should not wait for any other event.

Signed-off-by: Ricardo Neri <ricardo.neri@ti.com>