OMAPDSS: HDMI: Relocate N/CTS calculations into generic HDMI driver
The N and CTS parameters are relevant to all HDMI implmentations and
not specific to a given IP. Hence, the calculation is relocated
into the generic HDMI driver.

Also, deep color is removed from the calculation until such feature
is supported by the HDMI driver. This is to avoid direct IP register
readings from the generic HDMI driver.

Signed-off-by: Ricardo Neri <ricardo.neri@ti.com>