OMAPDSS: HDMI: Relocate N/CTS calculations into generic HDMI driver
The N and CTS parameters are relevant to all HDMI implmentations and
not specific to a given IP. Hence, the calculation is relocated
into the generic HDMI driver.

Also, deep color is not queried but it is still considered in the
calculation of N. This is to be changed when deep color support is
implemented in the driver.

Signed-off-by: Ricardo Neri <ricardo.neri@ti.com>