OMAPDSS: HDMI: OMAP4: Add an audio configuration function
The generic HDMI driver does not need to know about the specific
settings of the IP. Hence, it passes the audio configuration
and the IP library parses such configuration and sets the IP
accordingly. This patch introduces an IP-specific audio configuration
function. Also the DMA, format and core config functions are not
exposed to the generic HDMI driver anymore.

Signed-off-by: Ricardo Neri <ricardo.neri@ti.com>