ARM: OMAP: Display: Configure DSI Mux pads for OMAP5
Add a function omap5_dsi_mux_pads which configures the CONTROL_DSIPHY register
to enable/disabling the DSI1 or DSI2 lanes.

The CONTROL_DSIPHY register on OMAP4 was used for enabling/disabling the lanes
and configuring them as pull up or pull down. On OMAP5, this register is only
used to enable the lanes. There are separate registers to configure
pull up and down.

Use raw register writes to configure the CONTROL_DSIPHY registers since mux
isn't completely available for OMAP5 yet.

Signed-off-by: Archit Taneja <archit@ti.com>
Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>