OMAPDSS: DSI: Allow variable CLKIN divider in DSI PHY
On OMAP3 and OMAP4, the DSI PHY(DDR clock) output was the DSI PLL output divided
by 4. On OMAP5, the default value of this divider is 2, there is a feature of
having the divider as 4 by disabling the Duty Cycle Corrector, but it isn't
power efficient and requires setting of a few extra bits.

Introduce changes to not have a default CLKIN4DDR clock, make it a variable
clock and get the divider using dss_features.

Signed-off-by: Archit Taneja <archit@ti.com>
Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>