OMAPDSS: Limit DISPC sync signal and clock to HDMI when no VENC is supported
The DISPC sync signal and clock (if selected) for the TV output may come from
HDMI or the video encoder. This depends of the device. If the video encoder
is not supported by the device (as in OMAP5) return the HDMI clock selection.
If both are supported (as in OMAP4), read the value from the register.

Signed-off-by: Ricardo Neri <ricardo.neri@ti.com>