OMAPDSS:HDMI: Improve logic of signal selection for the DISPC
The display controller may take its sync signals from the video encoder
or HDMI. This patch checks if the selected source is supported.
Also, if the video encoder is not supported, there is no need to
set the source as HDMI is the only source. In such case reading/writing
DSS_CONTROL[15] is not possible.

While here, the error handling of the current caller is implemented.

Signed-off-by: Ricardo Neri <ricardo.neri@ti.com>