arm/dts: OMAP4: Add nodes to represent HDMI sound
The 'sound_hdmi' is similar to that of abe-twl6040 node.

TODO: find a better place for the codec. Perhaps it can be a child of the level
shifter or it can be added to the 'hdmi_audio' node. If it does not represent real
hardware, maybe just create the device from the machine driver. We still need an entry
in the DT, though.

Signed-off-by: Ricardo Neri <ricardo.neri@ti.com>