OMAPDSS: HDMI: Modify comment on actions up timings change.
When chaning the timings, we need to reconfigure audio. For this
we have two options either 1) reconfigure internally without
letting know the users or 2) notify the users and let them
react accordingly. 1) makes more sense as the users does not
need to know the internals of the reconfiguration and audio
will be restored to whatever state it had before the reconfiguration.

This implementation handles the changes internally. For this,
we keep a copy of the last configuration and use it to restore it
in the event of timings change. Audio is restored up to the point
it was before the timings change.

TODO: Evaluate what happens if for instance, timings are readjusted
while configuring. Should we abort and begin. Does the state machine
warraties good behavior? Perhaps use helper funcions to do the stuff
and make the current functions just wrappers that hold the locks.

Signed-off-by: Ricardo Neri <ricardo.neri@ti.com>