ASoC: OMAP: HDMI: Stop audio stream when the display becomes inactive
If there is an active stream, stop it if the display becomes inactive.

TODO: Reconsider if stopping the stream is a smart approach. Maybe
it is better to notify the userspace about the display going inactive
so that the user space may take action.

TODO: As the display goes inactive, DMA stalls. Check if ALSA is still
able to stop the stream from the user space.

Signed-off-by: Ricardo Neri <ricardo.neri@ti.com>