ARM: OMAP2+: HDMI: Refine the DMA resource for audio drivers
Instead of passing the complete resource to the audio drivers, just
pass the adress offset of the DMA port for audio samples. Thus,
we prevent that two drivers try to ioremap the same resources. This
is to be safe, as the audio drivers will not need to ioremap such
resource.

Signed-off-by: Ricardo Neri <rneri@dextratech.com>